requirements.txt
Pillow
tk
numpy
